[#2] - Basic generator in matlab
This generator attempts to model restaurant order by tables with the
number of people per table and the price of the bill for the whole table

This could be improved because a single table may wish to use multiple
cards to pay, so splitting it up like this may not be a realistic
option. Still, it gives us something to work with and refine.